Dereference symbolic links when copying files from ct data folder
When running `rebar3 ct`, `rebar_prv_common_test` copies the data folders
to a different location. If `{name}_SUITE_data` folder contains relative
symbolic links, the symbolic links might stop working since the path
might be different.

A solution is to add the `-L` option to `cp` so it dereferences the
symbolic links and copies the original content in the target file.

This is only done for the data folder but and won't change any code that
previously used `rebar_file_utils:cp_r`.