Do not dereference symbolic links in win32 when using cp_r
The current behaviour:
- When `Source` is a folder, `robocopy` is used. It follows symbolic
  links and copies the original data, creating a raw file

- When `Source` is a file/symbolic link, `file:copy` is used. It copies
  the raw data creating a raw file.

The change in this commit will use `/sl`  when copying files
from Source folder and `file:make_symlink` when Source is a
symbolic link on Windows.

If the `[{dereference, true}]` option is given, it will dereference the
files and copy the content instead of a symbolic link

This change has been added so the Windows copy behaves more unix one.